+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add menu action slots to manage plugins page
 - Add a share menu action for exporting Asset Link plugins
+- Add an installation and upgrades section to the docs
 
 ### Changed