+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Fix a bug with nested AND/OR query groups not being added as memberOf eachother
+- Fix a bug with the UrlBasedAssetSearcher plugin
 
 ## [1.0.0-alpha6] - 2023-03-17