+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Implement nested relationship attribute filtering
 - Surface navigation for exact route matches above asset search results
+- Surface option to install plugin and plugin list urls above asset search results
 
 ### Fixed